Integrating with Amazon RDS database
Adding database.go file defining a function that establishes a connection with a remote database
given an endpoint and password defined through environment variables.

Updating application.go to include a handler on the "/counter" route that serves up the latest values
of an integer counter and then increments the value in the database.

We are using password authentication to connect to the database instance because IAM authentication
is not available for the small instance size that we have provisioned in AWS. IAM authentication will also
not be supported if and when we switch to using a serverless deployment of Aurora, so we may need to continue
connecting in this way. Environment variables are used to store the endpoint and password so that we can configure
different environments to connect to different instances of the database and to ensure that we do not check in
database credentials to source control.