Adding automatic database migrations
Using golang-migrate to automatically execute database migrations on startup.
When we need to make a change to the structure of the database, we create a new
migration. Then, when we start the service, we will attempt to apply the migration
before anything else, so that we know it is in the expected state. Thus in most cases
we avoid the need to manually execute SQL queries or import data.

The initial migration creates the Counters table and inserts the row with
ID=1 that is expected by our existing API.

Migrations are defined in the 'migrations' folder, which is read at runtime and
must be packed along with the application binary as part of the CI/CD deployment.
Each migration defines a change made to the database (such as creating a
table or adding a column). For each migration we provide an "up" file for
making the change, and a corresponding "down" file for safely rolling it
back.
For more information, see: https://github.com/golang-migrate/migrate