Bug: .env files excluded from deployed package (#9)
Updating 'deploy' workflow to include .env* files.
Updating 'deploy' workflow to pack application binary in the root of the
zip so that its location is consistent relative to the .env* files in
the local and deployed environments.

We weren't actually packing and deploying the .env* files, so they were unable for read in the deployed environment. This meant that when I removed the manually-configured environment variables in the dev environment, the '/counters' endpoint would fail because the service was unable to read from the database.

I've updated the github actions workflow to pack files matching .env*, and I've also added error handling logic so that the service will immediately fail if we are unable to load the expected environment configuration files.