Adding tasks to start mysql-server in Docker and local DB configuration (#10)
Adding VSCode tasks to start or kill/remove a local mysql-server instance running in Docker.  

Adding new 'local' environment configuration and updating environment.go to make this the new default in the absence of an explicitly defined MATTERHORN_ENV.  

With these changes we will be able to easily spin up a local database running in a Docker container and use it for local development and testing.