Implementing APIs for Proof of Concept (#17)
Adding Swagger UI and implementing PoC APIs.

Updating controller logic to return BadRequest (400) instead of InternalServerError (500) when failing to parse required path parameters.

Updating service interfaces to optionally return status codes and pass those back to the controllers, which uses it in the response if one is provided. Adding a helper method to the database package to attempt to parse a specific MySQL error code from an error returned by the mysql database provider so that we can map errors to the appropriate HTTP status code for our APIs.

These are changes to generated code, which means that now we can't regenerate those file. They have been added to .openapi-generator-ignore to account for this. It would be better if we could automatically regenerate these files without losing our implementation in the event that we change interfaces or simply want to use updated generated code in the future, but it's more important to be able to return the correct status codes. We will still get value from the generator for bootstrapping the addition of new APIs.

Moving database and environment code to new database and environment packages.

Added unit tests for success cases on all APIs.