Adding database migration for adding test user data to database
Adding database migration that adds a row of test data to the Users database.
This migration also recreates the foreign key constraints on existing database tables
to define 'CASCADE' behavior in response to an update or delete of the row associated
with the foreign key.
This means that if a User row is deleted, then all rows in other tables with
a foreign key relationship to that User are also deleted.
If a user row's ID is updated, then all rows in other tables that currently map
to that row will have their foreign key values updated to maintain the relationship.
Previously, these updates did not "cascade", which meant that attempts to delete
a row from the Users table would simply fail if it would result in a violated foreign
key relationship (data associated with the User row in another table would have
foreign key values corresponding to nothing).