Requiring authentication for user data queries and mutations (#29)
Adding 'me' query endpoint that retrieves a User associated with the provided authentication token

Requiring authentication for most queries/mutations and limiting access
to only the authenticated user's data. User ID is taken from the
caller's token instead of being provided as input.

The exceptions are the 'user' query and 'createUser' mutation which
remain for development and testing purposes; these must be restricted input
Production environments before launch.

Adding environment variable configuration to allow developers to
manually configure a test user identity in non-production environments
to simplify testing. Rather than creating a test user and logging in to
get an access token, token validation is bypassed and the user ID may be
read from an environment variable, though it must begin with 'test|'.
(Note: requests will still fail if an INVALID token is provided, even in
a non-production environment; but requests without a token will succeed
if test data is configured for use instead.)